+++ src/Ema/Route/Slug.hs
@@ -1,14 +1,17 @@
+{-# LANGUAGE DeriveDataTypeable #-}
+{-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE InstanceSigs #-}
 
 module Ema.Route.Slug where
 
+import Data.Data (Data)
 import qualified Data.Text as T
 import qualified Data.Text.Normalize as UT
 import qualified Network.URI.Encode as UE
 
 -- | An URL path is made of multiple slugs, separated by '/'
 newtype Slug = Slug {unSlug :: Text}
-  deriving (Eq, Show, Ord)
+  deriving (Eq, Show, Ord, Data, Generic)
 
 -- | Decode an URL component into a `Slug` using `Network.URI.Encode`
 decodeSlug :: Text -> Slug